+++ src/features/remakeissue/constants/newsCategoryMapping.ts
@@ -1,0 +1,11 @@
+type CategoryMapping = {
+  [key: string]: string;
+};
+export const categoryMapping: CategoryMapping = {
+  ENTERTAINMENTS: '연예',
+  POLITICS: '정치',
+  ECONOMY: '경제',
+  SOCIETY: '사회',
+  WORLD: '국제',
+  SPORTS: '스포츠',
+};
