+++ src/features/remakeissue/remotes/sameCategoryReprocessedIssue.ts
@@ -1,8 +1,8 @@
 import { client } from '../../../shared/remotes/axios';
-import { SameCategoryReProcessedIssue } from '../../../shared/types/news';
+import { SameCategoryReprocessedIssue } from '../../../shared/types/news';
 
 export const getSameCategoryReprocessedIssues = async (current: number, category: string) => {
-  const { data } = await client.get<SameCategoryReProcessedIssue[]>(
+  const { data } = await client.get<SameCategoryReprocessedIssue[]>(
     '/reprocessed-issue/by-category',
     {
       params: { current: current, category: category },
